requirements.txt
# Core
timm==1.0.19

# Diffusion & HF stack
diffusers
transformers
huggingface-hub
safetensors

# Array / image / metrics / viz
numpy
Pillow
tqdm
matplotlib
scikit-image
opencv-python-headless
einops

# Losses
pytorch-ssim